Add StateLayerDashboard (Level 1) - The State assessment layer
- Created INFORMATION_ARCHITECTURE.md defining Top-Middle-Bottom 3-level system:
  * Level 0 (The Infinite): TimeHorizonDashboard - long-term context & milestones
  * Level 1 (The State): StateLayerDashboard - health/leverage/skills gauges
  * Level 2 (The Immediate): TodayInfoCard + ProjectCards + apps - daily actions

- Created StateLayerDashboard.kt skeleton with:
  * ArcGauge component (circular progress + trend indicator)
  * Placeholders for HealthGauge, SuperTrajectory, SkillsDashboard
  * Phase 1 roadmap: Health + Time Allocation gauges
  * Phase 2 roadmap: Super/Leverage + Skills + interactive detail sheets

- Integrated StateLayerDashboard into HomePage between Level 0 and Level 2
- Establishes complete information hierarchy: orient → assess → execute

+++ app/src/main/java/com/ambient/launcher/home/TimeHorizonDashboard.kt
@@ -443,7 +443,7 @@ private fun getSeasonColor(season: String): Color {
 }
 
 private fun getWeekOfYear(date: LocalDate): Int {
-    return java.time.temporal.WeekFields.ISO.weekOfYear.getFrom(date)
+    return java.time.temporal.WeekFields.ISO.weekOfYear().getFrom(date).toInt()
 }
 
 private fun calculateLifeProgress(): LifeProgress {
